# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sree Ram Technologies management system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /installations:
    get:
      operationId: listInstallations
      tags: [installations]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Installation"
    post:
      operationId: createInstallation
      tags: [installations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installation"
  /installations/{id}:
    patch:
      operationId: updateInstallation
      tags: [installations]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallationInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installation"
    delete:
      operationId: deleteInstallation
      tags: [installations]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    patch:
      operationId: updateTicket
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    delete:
      operationId: deleteTicket
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Installation:
      type: object
      required: [id, customerId, customerName, siteName, address, installedDate, status, totalCameras, amcExpiry, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: string }
        siteName: { type: string }
        address: { type: string }
        installedDate: { type: string, format: date }
        status:
          type: string
          enum: [active, maintenance, decommissioned]
        totalCameras: { type: integer }
        amcExpiry: { type: string, format: date, nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    InstallationInput:
      type: object
      required: [customerId, siteName, address, installedDate, status, totalCameras]
      properties:
        customerId: { type: integer }
        siteName: { type: string, minLength: 1 }
        address: { type: string, minLength: 1 }
        installedDate: { type: string, format: date }
        status:
          type: string
          enum: [active, maintenance, decommissioned]
        totalCameras: { type: integer, minimum: 0 }
        amcExpiry: { type: string, format: date, nullable: true }
        notes: { type: string, nullable: true }
    Ticket:
      type: object
      required: [id, ticketNumber, customerId, customerName, subject, description, priority, status, category, createdAt]
      properties:
        id: { type: integer }
        ticketNumber: { type: string }
        customerId: { type: integer }
        customerName: { type: string }
        subject: { type: string }
        description: { type: string }
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, on_hold, resolved, closed]
        category:
          type: string
          enum: [installation, repair, amc, complaint, query, replacement]
        assignedToId: { type: integer, nullable: true }
        assignedToName: { type: string, nullable: true }
        resolvedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    TicketInput:
      type: object
      required: [customerId, subject, description, priority, status, category]
      properties:
        customerId: { type: integer }
        subject: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, on_hold, resolved, closed]
        category:
          type: string
          enum: [installation, repair, amc, complaint, query, replacement]
        assignedToId: { type: integer, nullable: true }
    Invoice:
      type: object
      required: [id, invoiceNumber, customerId, customerName, issueDate, dueDate, status, subtotal, tax, total, createdAt]
      properties:
        id: { type: integer }
        invoiceNumber: { type: string }
        customerId: { type: integer }
        customerName: { type: string }
        issueDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        subtotal: { type: number }
        tax: { type: number }
        total: { type: number }
        notes: { type: string, nullable: true }
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        createdAt: { type: string, format: date-time }
    InvoiceItem:
      type: object
      required: [description, quantity, unitPrice, amount]
      properties:
        description: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        amount: { type: number }
    InvoiceInput:
      type: object
      required: [customerId, issueDate, dueDate, status, subtotal, tax, total, items]
      properties:
        customerId: { type: integer }
        issueDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        subtotal: { type: number, minimum: 0 }
        tax: { type: number, minimum: 0 }
        total: { type: number, minimum: 0 }
        notes: { type: string, nullable: true }
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
...